Moved Navbar to index and removed its references. Also moved it's subcomponents into their own subfolder
