+++ src/App.js
@@ -1,12 +1,10 @@
 import React, { Component } from 'react';
-import Navbar from './components/Navbar.js';
 
 class App extends Component {
   render() {
     return (
       <div className="App">
-        <Navbar />
-        <p style={{alignItems: 'center'}}>Hello World</p>
+        <p style={{alignItems: 'center', justifyContent: 'center'}}>Welcome to Opencourse!</p>
       </div>
     );
   }
