+++ src/Home.js
@@ -1,5 +1,4 @@
 import React, { Component } from 'react';
-import Navbar from './components/Navbar.js';
 import Launch from './pages/Launch.js';
 
 export default class Home extends Component {
@@ -7,7 +6,6 @@ export default class Home extends Component {
   render() {
     return (
       <div>
-        <Navbar />
         <Launch />
       </div>
     );
