+++ src/index.js
@@ -7,6 +7,7 @@ import {
 } from 'react-router-dom';
 
 // Components
+import Navbar from './components/Navbar.js';
 import Home from './Home.js';
 import App from './App.js';
 import Login from './pages/Login.js';
@@ -23,6 +24,7 @@ import 'bootstrap/dist/css/bootstrap.min.css';
 
 ReactDOM.render(
   <FirebaseContext.Provider value={new Firebase()}>
+  <Navbar />
   <Router>
     <Route exact path='/' component={Home}/>
     <Route path='/hello' component={App}/>
@@ -30,7 +32,7 @@ ReactDOM.render(
     <Route exact path='/firebase' component={Firebase}/>
     <Route path='/home' component={UserHome}/>
     <Route exact path='/newcourse' component={NewCourse}/>
-    <Route exact path='/editcourseview' component={EditCourse}/>
+    <Route path='/editcourseview/' component={EditCourse}/>
     <Route exact path='/courses' component={Courses}/>
     <Route path='/course/' component={CourseView}/>
   </Router>
