+++ src/pages/Login.js
@@ -1,5 +1,4 @@
 import React, { Component } from 'react';
-import NavBar from '../components/Navbar.js';
 
 import '../components/Firebase/firebaseAuth.js';
 import 'firebaseui/dist/firebaseui.css';
@@ -8,7 +7,6 @@ export default class Login extends Component {
   render(){
     return(
       <div>
-      <NavBar />
       <div style={jumbotronStyle}>
         <h1>Login</h1>
         <h4>to get started with your new learning adventure</h4>
